coding-problems/src/main/java/com/problems/median/RunningMedian.java
package com.problems.median;

import java.util.Collections;
import java.util.PriorityQueue;

/**
 * Created by chinmay on 5/4/2017.
 */
public class RunningMedian {

    private PriorityQueue<Integer> small = new PriorityQueue<>(Collections.reverseOrder());
    private PriorityQueue<Integer> large = new PriorityQueue<>();
    private boolean even = true;

    public double findMedian() {
        if (even)
            return (small.peek() + large.peek()) / 2.0;
        else
            return small.peek();
    }

    public void add(int num) {
        if (even) {
            large.offer(num);
            small.offer(large.poll());
        } else {
            small.offer(num);
            large.offer(small.poll());
        }
        even = !even;
    }


}
